Fix power range

+++ src/qibocal/protocols/characterization/readout_optimization/twpa_calibration/power.py
@@ -59,8 +59,7 @@ def _acquisition(
 
     power_range = np.arange(
         -params.power_width / 2, params.power_width / 2, params.power_step
-    ).astype(int)
-
+    )
     data = TwpaPowerData()
 
     initial_twpa_power = {}
